[#24]ref: monitor.c, monitor_utils.c: adjust to norm
