+++ philo/src/thread/monitor.c
@@ -28,58 +28,13 @@ static void	*set_finish(t_data *data)
 	return (NULL);
 }
 
-void	set_all_philo_dead(t_data *data)
-{
-	size_t	i;
-
-	i = 0;
-	while (i < (size_t)data->num_of_philo)
-		data->philos[i++]->is_dead = TRUE;
-}
-
-void	all_philos_mutex_lock(t_data *data, size_t num)
-{
-	size_t	i;
-
-	i = 0;
-	while (i < (size_t)data->num_of_philo)
-	{
-		if (i != num)
-			pthread_mutex_lock(&(data->count_mtxs[i]->mtx));
-		i++;
-	}
-}
-
-void	all_philos_mutex_unlock(t_data *data)
-{
-	size_t	i;
-
-	i = 0;
-	while (i < (size_t)data->num_of_philo)
-	{
-		pthread_mutex_unlock(&(data->count_mtxs[i]->mtx));
-		i++;
-	}
-}
-
-void	died(t_data *data, size_t i, long long now)
-{
-	pthread_mutex_lock(&(data->msg_mtx->mtx));
-	pthread_mutex_lock(&(data->flag_mtx->mtx));
-	printf(RED);
-	printf("%lld %zu died\n", now - data->start_time, i);
-	printf(RESET);
-	data->is_finish = TRUE;
-	pthread_mutex_unlock(&(data->flag_mtx->mtx));
-	pthread_mutex_unlock(&(data->msg_mtx->mtx));
-}
-
 static t_bool	is_philo_dead(t_data *data, size_t i)
 {
 	long long	now;
 
 	now = get_mtime();
-	if (data->philos[i]->last_eat && now - data->philos[i]->last_eat >= data->philos[i]->time_to_die)
+	if (data->philos[i]->last_eat && \
+		now - data->philos[i]->last_eat >= data->philos[i]->time_to_die)
 	{
 		all_philos_mutex_lock(data, i);
 		set_all_philo_dead(data);
@@ -109,7 +64,8 @@ void	*monitor(void *arg)
 			pthread_mutex_lock(&(data->count_mtxs[i]->mtx));
 			if (is_philo_dead(data, i) == TRUE)
 				return (NULL);
-			if (data->must_eat > 0 && data->philos[i]->eat_count >= data->must_eat)
+			if (data->must_eat > 0 && \
+				data->philos[i]->eat_count >= data->must_eat)
 				totla_eat++;
 			pthread_mutex_unlock(&(data->count_mtxs[i]->mtx));
 			i++;
